Fix excess newlines being added to pages when navigated to
We've been triggering shortcuts on keydown - but often were completely
rendering the new page and focusing the body before the corresponding
keyup. So, when you hit enter it was inserted into the destination
page's body.
All we need to do to fix it is call event.preventDefault() in the hotkey
responder.